requirements.txt
Flask==2.2.0
werkzeug==2.1.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
python-binance==1.0.29
